Added HTTP headers to help link queries to clicks
The API now returns an X-Correlation-Id header from search requests,
which SDK clients are presumed to return with subsequent resolution
requests. This allows better performance tracking and tuning when
serving relevant content.

Search Requests can now pass placement tags which will allow for
labeling where in the app a query is occurring, and to study the
performance by tag or groups of tags.

If the AdSupport framework is included in the app which includes
URXSearch, the IDFA will also be passed in the HTTP requests.